Refactor Cosmos components
