+++ src/app/cosmos/[fixture]/nextShared.ts
@@ -1,0 +1,3 @@
+export type CosmosPageParams = {
+  fixture: string | null;
+};
